+++ zbus/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Re-export `logcontrol` and `logcontrol::DBUS_OBJ_PATH`.
+- Add `logcontrol_zbus::ConnectionBuilderExt` to extend `ConnectionBuilder` with `serve_log_control`.
 
 ## [1.0.0] – 2023-09-30